renamed event_search to events
