+++ metecho/__init__.py
@@ -3,5 +3,6 @@ from .version import __version__
 from . import tools
 from . import data
 from . import plot
+from . import events
 
 from .tools import PROFILER as profiler
